# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ennote ennoworker API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /v1/graphs:
    get:
      operationId: listGraphs
      summary: List Worker-global Graph source summaries without loading private resources
      responses:
        "200":
          description: Global Graph catalog
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { type: array, items: { $ref: "#/components/schemas/GraphSummary" } } }
    post:
      operationId: createGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [id, name]
              properties:
                id: { type: string, pattern: "^[a-z][a-z0-9-]{1,63}$" }
                name: { type: string, minLength: 1 }
      responses:
        "201":
          description: Empty editable Graph draft created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphDetail" } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/graphs/{graphID}:
    parameters:
      - { name: graphID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getGraph
      responses:
        "200":
          description: Parsed Graph draft
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphDetail" } }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateGraph
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GraphMutation" }
      responses:
        "200":
          description: Semantically updated Graph draft
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GraphDetail" } }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /v1/provider-profiles/discover-models:
    post:
      operationId: discoverProviderModels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                providerId: { type: string }
                baseUrl: { type: string, format: uri }
                apiKey: { type: string }
      responses:
        "200":
          description: Fetched model catalog
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/DiscoveredModel" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
...
  /v1/provider-profiles/{providerID}/test:
    parameters:
      - name: providerID
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: testProviderProfile
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                modelProfileId: { type: string }
      responses:
        "200":
          description: Ephemeral Provider diagnostic result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ProviderDiagnostic" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
...
    DiscoveredModel:
      type: object
      required: [modelName]
      additionalProperties: false
      properties:
        modelName: { type: string }
        displayName: { type: string }
        contextWindow: { type: integer, minimum: 0 }
        maxOutputTokens: { type: integer, minimum: 0 }
        supportsVision: { type: boolean }
        supportsThinking: { type: boolean }
    ProviderDiagnosticStage:
      type: object
      required: [name, status, message, latencyMs]
      properties:
        name: { type: string, enum: [configuration, credentials, model, generation] }
        status: { type: string, enum: [passed, failed] }
        message: { type: string }
        latencyMs: { type: integer, format: int64, minimum: 0 }
    ProviderFailure:
      type: object
      required: [category, message, retryable]
      properties:
        category:
          type: string
          enum:
            - configuration_invalid
            - credential_unavailable
            - endpoint_unreachable
            - authentication_failed
            - model_not_found
            - rate_limited
            - request_timeout
            - context_overflow
            - malformed_response
            - provider_internal_error
            - request_rejected
            - cancelled
            - unknown
        message: { type: string }
        retryable: { type: boolean }
        requestId: { type: string }
    ProviderDiagnostic:
      type: object
      required: [providerId, status, stages, latencyMs, testedAt]
      properties:
        providerId: { type: string }
        modelProfileId: { type: string }
        modelName: { type: string }
        status: { type: string, enum: [ready, failed] }
        failure: { $ref: "#/components/schemas/ProviderFailure" }
        stages:
          type: array
          items: { $ref: "#/components/schemas/ProviderDiagnosticStage" }
        latencyMs: { type: integer, format: int64, minimum: 0 }
        testedAt: { type: string, format: date-time }
    ThinkingEffort:
      type: string
      enum: [default, low, medium, high]
...
    GraphTask:
      type: object
      additionalProperties: false
      required: [name, goal]
      description: Role-backed Tasks declare role only; inline Tasks declare model with optional thinking and skills. The modes are mutually exclusive and enforced by the Worker.
      properties:
        name: { type: string, minLength: 1 }
        role: { type: string, pattern: "^(local|global)/[a-z][a-z0-9_-]{1,63}$" }
        model: { type: string, pattern: "^[^/]+/[^/]+$" }
        thinking: { $ref: "#/components/schemas/ThinkingEffort" }
        skills: { type: array, uniqueItems: true, items: { type: string, pattern: "^(local|global)/[a-z][a-z0-9_-]{1,63}$" } }
        goal: { type: string, minLength: 1 }
        writes: { type: array, uniqueItems: true, items: { type: string } }
        budget:
          type: object
          additionalProperties: false
          properties: { tokens: { type: integer, minimum: 1 } }
    GraphDocument:
      type: object
      additionalProperties: false
      required: [schemaVersion, id, name, tasks, graph]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        id: { type: string }
        name: { type: string }
        description: { type: string }
        tasks: { type: object, additionalProperties: { $ref: "#/components/schemas/GraphTask" } }
        graph:
          type: object
          additionalProperties: { type: array, uniqueItems: true, items: { type: string } }
    GraphSummary:
      type: object
      additionalProperties: false
      required: [id, name, path, latestVersion]
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        digest: { type: string }
        error: { type: string }
        latestVersion: { type: integer, minimum: 0 }
    GraphDetail:
      allOf:
        - $ref: "#/components/schemas/GraphSummary"
        - type: object
          required: [digest, document]
          properties:
            digest: { type: string }
            document: { $ref: "#/components/schemas/GraphDocument" }
    GraphMutation:
      type: object
      additionalProperties: false
      required: [expectedDigest]
      properties:
        expectedDigest: { type: string }
        name: { type: string }
        description: { type: string }
        task:
          type: object
          additionalProperties: false
          required: [id]
          properties:
            id: { type: string }
            value: { $ref: "#/components/schemas/GraphTask" }
        dependencies:
          type: object
          additionalProperties: false
          required: [taskId, depends]
          properties:
            taskId: { type: string }
            depends: { type: array, uniqueItems: true, items: { type: string } }
...